Slack-notice/requirements.txt
# Function dependencies, for example:
# package>=version


requests>=2.27.1
pandas>=1.2.4
numpy>=1.19.5
yfinance>=0.1.70 
slackclient>=2.5.0


